Minor optimizations: Switched BatchPlanner.BatchPlan to try to use ArrayLists instead of HashSets whenever it can. Cleaned up LocalTransaction source. Assume that each PlanFragment can return one and only one output VoltTable

+++ src/frontend/edu/brown/hstore/PartitionExecutor.java
@@ -1874,7 +1874,7 @@ public class PartitionExecutor implements Runnable, Shutdownable, Loggable {
                             output_depIds,
                             parameterSets,
                             batchSize,
-                            txn_id,
+                            txn_id.longValue(),
                             this.lastCommittedTxnId,
                             undoToken);
